added user details update feature added user account delete feature fixed some bugs now bike details will br auto updated after booking or return in UI jwt expire duration set to 1d

+++ Backend/controllers/authController.ts
@@ -3,7 +3,7 @@ import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
 import User, { IUser } from '../models/user';
 import envVars from '../config/config';
-import { Types } from 'mongoose';
+import Booking from '../models/booking';
 
 // Handle user registration
 export const register = async (req: Request, res: Response) => {
@@ -53,9 +53,8 @@ export const login = async (req: Request, res: Response) => {
         if (!isMatch) {
             return res.status(400).json({ message: 'Invalid credentials' });
         }
-
         // Create and send token
-        const token = jwt.sign({ id: user._id }, envVars.jwtSecret, { expiresIn: '1h' });
+        const token = jwt.sign({ id: user._id }, envVars.jwtSecret, { expiresIn: '1d' });
         res.json({ token });
     } catch (error) {
         res.status(500).json({ message: 'Server error', error });
@@ -77,14 +76,26 @@ export const verifyEmail = async (req: Request, res: Response) => {
     // TODO: Implement email verification logic
 };
 
-export const deleteUser = async (id: Types.ObjectId, req: Request, res: Response) => {
+export const deleteUser = async (req: Request, res: Response) => {
     try {
-        const user: IUser | null = await User.findById(id);
+        const id = req.body.user.id;
+        const { password } = req.body;
 
+        const user: IUser | null = await User.findById(id);
         if (!user) {
             return res.status(404).json({ message: 'User not found' });
         }
 
+        // Compare password
+        const isMatch = await bcrypt.compare(password, user.password);
+        if (!isMatch) {
+            return res.status(400).json({ message: 'Invalid credentials' });
+        }
+
+        // Delete user's bookings
+        await Booking.deleteMany({ userId: id });
+
+        // Delete the user
         await User.findByIdAndDelete(id);
 
         res.json({ message: 'User deleted successfully' });
